Rename src to google_places_enricher_2_0 and update imports
Rename the src directory to google_places_enricher_2_0 so we can use absolute imports. Update all internal imports to use the new package path. Add run_app.py for ease of running with python run_app.py.
