+++ google_places_enricher_2_0/static/coordinates_definition.css
