+++ google_places_enricher_2_0/templates/categories.html
