+++ google_places_enricher_2_0/templates/coordinates_definition.html
