+++ google_places_enricher_2_0/templates/enrich_data.html
