+++ google_places_enricher_2_0/templates/view_data.html
