+++ run_app.py
@@ -1,0 +1,4 @@
+from google_places_enricher_2_0.app import app
+
+if __name__ == "__main__":
+    app.run()
